+++ Release/Changelog.md
@@ -1,5 +1,4 @@
-## v2.6.2 (unreleased)
-- Update version number in Info.cs.
+## v2.6.2
 - Add handling for exceptions thrown while preloading bro assets.
 - Fix handling array of special icons as multiple variants rather than a single variant with multiple specials.